Factored out common code from AsyncUsbSender and AsyncUsbReceiver into AsycUsbTransceiverBase
