+++ plugins/usbdmx/AsyncUsbReceiver.cpp
@@ -26,42 +26,18 @@ namespace ola {
 namespace plugin {
 namespace usbdmx {
 
-namespace {
-
-/*
- * Called by libusb when the transfer completes.
- */
-#ifdef _WIN32
-__attribute__((__stdcall__))
-#endif
-void AsyncCallback(struct libusb_transfer *transfer) {
-  AsyncUsbReceiver *widget = reinterpret_cast<AsyncUsbReceiver*>(
-      transfer->user_data);
-  widget->TransferComplete(transfer);
-}
-}  // namespace
-
 AsyncUsbReceiver::AsyncUsbReceiver(ola::usb::LibUsbAdaptor *adaptor,
                                    libusb_device *usb_device,
                                    PluginAdaptor *plugin_adaptor)
-    : m_adaptor(adaptor),
-      m_usb_device(usb_device),
-      m_usb_handle(NULL),
+    : AsyncUsbTransceiverBase(adaptor, usb_device),
       m_plugin_adaptor(plugin_adaptor),
-      m_inited_with_handle(false),
-      m_suppress_continuation(false),
-      m_transfer_state(IDLE) {
-  m_transfer = m_adaptor->AllocTransfer(0);
-  m_adaptor->RefDevice(usb_device);
+      m_inited_with_handle(false) {
 }
 
 AsyncUsbReceiver::~AsyncUsbReceiver() {
-  CancelTransfer();
   if (!m_inited_with_handle) {
     m_adaptor->Close(m_usb_handle);
   }
-  m_adaptor->UnrefDevice(m_usb_device);
-  m_adaptor->FreeTransfer(m_transfer);
 }
 
 bool AsyncUsbReceiver::Init() {
@@ -83,66 +59,6 @@ bool AsyncUsbReceiver::Start() {
   }
   ola::thread::MutexLocker locker(&m_mutex);
   return PerformTransfer();
-}
-
-void AsyncUsbReceiver::CancelTransfer() {
-  if (!m_transfer) {
-    return;
-  }
-
-  bool canceled = false;
-  while (1) {
-    ola::thread::MutexLocker locker(&m_mutex);
-    if (m_transfer_state == IDLE || m_transfer_state == DISCONNECTED) {
-      break;
-    }
-    if (!canceled) {
-      m_suppress_continuation = true;
-      if (m_adaptor->CancelTransfer(m_transfer) == 0) {
-        canceled = true;
-      } else {
-        break;
-      }
-    }
-  }
-
-  m_suppress_continuation = false;
-}
-
-void AsyncUsbReceiver::FillControlTransfer(unsigned char *buffer,
-                                           unsigned int timeout) {
-  m_adaptor->FillControlTransfer(m_transfer, m_usb_handle, buffer,
-                                 &AsyncCallback, this, timeout);
-}
-
-void AsyncUsbReceiver::FillBulkTransfer(unsigned char endpoint,
-                                        unsigned char *buffer,
-                                        int length,
-                                        unsigned int timeout) {
-  m_adaptor->FillBulkTransfer(m_transfer, m_usb_handle, endpoint, buffer,
-                              length, &AsyncCallback, this, timeout);
-}
-
-void AsyncUsbReceiver::FillInterruptTransfer(unsigned char endpoint,
-                                             unsigned char *buffer,
-                                             int length,
-                                             unsigned int timeout) {
-  m_adaptor->FillInterruptTransfer(m_transfer, m_usb_handle, endpoint, buffer,
-                                   length, &AsyncCallback, this, timeout);
-}
-
-int AsyncUsbReceiver::SubmitTransfer() {
-  int ret = m_adaptor->SubmitTransfer(m_transfer);
-  if (ret) {
-    OLA_WARN << "libusb_submit_transfer returned "
-             << m_adaptor->ErrorCodeToString(ret);
-    if (ret == LIBUSB_ERROR_NO_DEVICE) {
-      m_transfer_state = DISCONNECTED;
-    }
-    return false;
-  }
-  m_transfer_state = IN_PROGRESS;
-  return ret;
 }
 
 void AsyncUsbReceiver::TransferComplete(struct libusb_transfer *transfer) {
